actual fix

+++ knn.h
@@ -7,6 +7,6 @@
 
 bool isPoison(input_Mushroom *checkMushrooms, data_Mushroom** knownMushrooms, int length);
 
-float euclidDistance(input_Mushroom *myMushroom, data_Mushroom *knownMushrooms);
+double euclidDistance(input_Mushroom *myMushroom, data_Mushroom *knownMushrooms);
 
 #endif
